requirements-prod.txt
boto3==1.37.28
click==8.1.7
fastapi==0.109.2
github-webhooks-framework2==0.2.2
ipwhois==1.3.0
jinja2-cli==0.8.2
netifaces==0.11.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.5
uvicorn==0.32.1
